back/requirements.txt
eth_account==0.11.0
Flask==3.0.2
flask_cors==4.0.0
hexbytes==1.0.0
python-dotenv==1.0.1
web3==6.15.1
